schedule_backend/pyproject.toml
[project]
name = "schedule-backend"
version = "0.1.0"
description = "AI 智能日程管理后端"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
    "openai>=1.57.0",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "pgvector>=0.3.0",
    "alembic>=1.14.0",
    "psycopg2-binary>=2.9.0",
]

[tool.uv]
# 国内网络环境可配置镜像源（超时可尝试切换）
# 清华: https://pypi.tuna.tsinghua.edu.cn/simple
# 阿里云: https://mirrors.aliyun.com/pypi/simple/
# 中科大: https://pypi.mirrors.ustc.edu.cn/simple/
index-url = "https://mirrors.aliyun.com/pypi/simple/"
